# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Complete eSIM platform with QPay payment integration and RoamWiFi API",
        "title": "eSIM Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "get": {
                "description": "Retrieve list of all available eSIM products with MNT pricing",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "parameters": [
                    {
                        "enum": [
                            "Asia",
                            "Europe",
                            "Africa",
                            "Americas",
                            "Oceania"
                        ],
                        "type": "string",
                        "description": "Filter by continent",
                        "name": "continent",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "true",
                            "false"
                        ],
                        "type": "string",
                        "description": "Filter by active status",
                        "name": "active",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products/continents": {
            "get": {
                "description": "Retrieve active products grouped by continent",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get products grouped by continent",
                "responses": {
                    "200": {
                        "description": "Products grouped by continent",
                        "schema": {
                            "$ref": "#/definitions/handlers.ProductsByContinentResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products/sku/{skuId}": {
            "get": {
                "description": "Public: Retrieve metadata for a specific SKU",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get single SKU info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SKU ID",
                        "name": "skuId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SKU details",
                        "schema": {
                            "$ref": "#/definitions/services.SKUInfo"
                        }
                    },
                    "404": {
                        "description": "SKU not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve SKU",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products/sku/{skuId}/packages": {
            "get": {
                "description": "Retrieve available data packages (plans) for a specific product SKU (day passes, data/validity variants)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get packages for a SKU",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SKU ID",
                        "name": "skuId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "If true returns detailed provider package structure",
                        "name": "detailed",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detailed packages with pricing when detailed=true",
                        "schema": {
                            "$ref": "#/definitions/services.EnrichedRoamWiFiPackagesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products/skus": {
            "get": {
                "description": "Public: Retrieve the raw list of available eSIM SKUs from provider (for client SKU selection)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get available SKUs",
                "responses": {
                    "200": {
                        "description": "List of SKUs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/services.SKUInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve SKUs",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Retrieve a single product by its UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "schema": {
                            "$ref": "#/definitions/handlers.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid product ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ProductResponse": {
            "type": "object",
            "properties": {
                "base_price": {
                    "type": "number"
                },
                "continent": {
                    "type": "string"
                },
                "countries": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "custom_price_usd": {
                    "type": "number"
                },
                "data_limit": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_price": {
                    "type": "number"
                },
                "exchange_rate": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_synced_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price_mnt": {
                    "type": "number"
                },
                "profit_margin": {
                    "type": "number"
                },
                "sku_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "validity_days": {
                    "type": "integer"
                }
            }
        },
        "handlers.ProductsByContinentResponse": {
            "type": "object",
            "properties": {
                "Africa": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ProductResponse"
                    }
                },
                "Asia": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ProductResponse"
                    }
                },
                "Europe": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ProductResponse"
                    }
                },
                "Global": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ProductResponse"
                    }
                },
                "North America": {
                    "description": "matches inferred continent value",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ProductResponse"
                    }
                },
                "Oceania": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ProductResponse"
                    }
                }
            }
        },
...
        "services.EnrichedRoamWiFiPackage": {
            "type": "object",
            "properties": {
                "api_code": {
                    "type": "string"
                },
                "days": {
                    "type": "integer"
                },
                "effective_price_mnt": {
                    "type": "number"
                },
                "effective_price_usd": {
                    "description": "Pricing enrichment",
                    "type": "number"
                },
                "expire_days": {
                    "type": "integer"
                },
                "flow_type": {
                    "type": "integer"
                },
                "flows": {
                    "type": "number"
                },
                "had_daypass_detail": {
                    "type": "integer"
                },
                "markup_percent": {
                    "type": "number"
                },
                "max_day": {
                    "type": "integer"
                },
                "max_discount": {
                    "type": "integer"
                },
                "min_day": {
                    "type": "integer"
                },
                "must_date": {
                    "type": "integer"
                },
                "network": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.RoamWiFiPackageNetwork"
                    }
                },
                "open_card_fee": {
                    "type": "number"
                },
                "overlay": {
                    "type": "integer"
                },
                "override_price_usd": {
                    "type": "number"
                },
                "pid": {
                    "type": "integer"
                },
                "premark": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "price_id": {
                    "type": "integer"
                },
                "price_source": {
                    "type": "string"
                },
                "show_name": {
                    "type": "string"
                },
                "single_discount": {
                    "type": "integer"
                },
                "single_discount_day": {
                    "type": "integer"
                },
                "support_daypass": {
                    "type": "integer"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "services.EnrichedRoamWiFiPackagesResponse": {
            "type": "object",
            "properties": {
                "country_code": {
                    "type": "string"
                },
                "country_images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.RoamWiFiCountryImage"
                    }
                },
                "display": {
                    "type": "string"
                },
                "display_en": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "packages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.EnrichedRoamWiFiPackage"
                    }
                },
                "sku_id": {
                    "type": "integer"
                },
                "support_country": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "services.RoamWiFiCountryImage": {
            "type": "object",
            "properties": {
                "country_code": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "name_en": {
                    "type": "string"
                }
            }
        },
        "services.RoamWiFiPackageNetwork": {
            "type": "object",
            "properties": {
                "namecn": {
                    "type": "string"
                },
                "nameen": {
                    "type": "string"
                },
                "operator": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "services.SKUInfo": {
            "type": "object",
            "properties": {
                "country_code": {
                    "type": "string"
                },
                "display": {
                    "type": "string"
                },
                "sku_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}